move transport2018 dataclasses to energy_demand
